Simulator modifications
Modification of simulator case 7 in order to avoid having any blocks in the simulator.
Modification of simulator case 8 in order to start to consider the dynamic system as an LPV System and not LTI System. Hence evolution of the paradigm in terms of controller tuning: the controller is adaptative to the change in velocity and some damping coefficient and pulsation of the loop can be designed in terms of profiles. Evolution of the EKF model to be considered to in the next commit.