H_Infinity standard synthesis using filters.
Major improvements in the use of the H Infinity synthesis to design a controller allowing a very low amplitude of commanded rudder angle. Therefore offering to avoid the limitation of the saturation cascade modelling the actuators. The system response if healthy inpresence of random wave signal input and very high constraints saturations on the input rudder angle.